fix OpenGL pasta rendering
Using the `Length` property of the `BufferObject` for the `count` parameter in the `DrawElements` call was the issue. The `Length` was more appropriately called '`ByteLength`', which is not what the `count` parameter was expecting (it expects total indices, not total byte length).

So, the `BufferObject` now has a `Length` *and* `ByteLength` property, and rendering is fixed.