remove small NativeMemoryPool functionality
It seemed that casting the `Memory<byte>` returned from the small `NativeMemoryManager` to the required `T` was actually taking longer than just instantiating a new `NativeMemoryManager<T>`. So I've removed the code path and related members for the small allocation, relying solely on creating a new `NativeMemoryManager<T>` for each rent call.