fix DrawElementsIndirectCommand incorrect count
DrawElementsIndirectCommand's `Count` parameter should be the number of *indexes* rendered, and not the number of vertexes. Rendering by the amount of vertexes was causing many tris to be left out of the rendering.

Also did some code cleanup from Rider auto-shortening all of my declarations from `new MyType()` to `new()`. Not a huge fan of that new syntax.