remove EntityManager enumerable caching
The enumerable caching was mostly a temporary solution that could later be expanded to a full caching solution, but for now I'm simply going to rely on the speed and efficiency of the implemented `Entity.Enumerator` struct. If it proves to not be enough in the future, I'll look into some other options.